started splitting txt2html.c up; added rule.c, txt2html.h.
Should've done this a while ago but I was being stubborn (thinking it
would be easier in a single file).

txt2html.h has a NodeTypes enum with the previously #define'd values.

I've removed the #define RULE_ because it was ugly and a little
dangerous, doing it in functions allows me to make sure str is long
enough before running the checks. It also makes the code cleaner since
everything is a combination of NodeTypes.